Refactor: Input 공용 컴포넌트 리팩토링
