--- src/shared/components/atoms/input/DefaultInputLine.vue
@@ -1,13 +1,0 @@
-<script setup>
-import BaseInput from './BaseInput.vue';
-
-const emit = defineEmits(['update:modelValue']);
-</script>
-
-<template>
-  <BaseInput
-    variant="line-default"
-    v-bind="$attrs"
-    @update:model-value="emit('update:modelValue', $event)"
-  />
-</template>
