--- src/shared/components/atoms/input/SubInputLine.vue
@@ -1,13 +1,0 @@
-<script setup>
-import BaseInput from './BaseInput.vue';
-
-const emit = defineEmits(['update:modelValue']);
-</script>
-
-<template>
-  <BaseInput
-    variant="line-sub"
-    v-bind="$attrs"
-    @update:model-value="emit('update:modelValue', $event)"
-  />
-</template>
